# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Orion Belt API
  description: |
    REST and WebSocket API for the Orion Belt privileged access gateway.

    **Base URL:** `/api/v1` (plus `/health` and optional `/metrics` outside the versioned prefix).

    **Authentication** (protected routes — try in order):
    1. `X-API-Key` header
    2. `X-Session-Token` header, `session_token` cookie, or `token` / `access_token` query (WebSocket)
    3. `Authorization: Bearer <JWT>`

    Pubkey login (`/public/login*`) requires a prior `POST /public/auth/challenge` and a signature proving possession of the private key.
    Password login: `POST /public/login/password` (password + TOTP). Console also supports WebAuthn and `osh login` browser bootstrap.

    Optional **SSH CA** endpoints issue short-lived User certs and expose Host CA trust material (`docs/SSH_CA.md`).

    **Admin routes** (`/api/v1/admin/*`) additionally require role `admin` or `operator` (or `is_admin`).

    **Rate limit:** protected (non-admin) routes — default 600 req/min per user (`auth.rate_limit_per_minute`).
  version: "0.8.1"
  contact:
    name: Orion Belt Maintainers
    url: https://github.com/zrougamed/orion-belt
  license:
    name: Apache-2.0 WITH Commons-Clause
    url: https://github.com/zrougamed/orion-belt/blob/master/LICENSE
servers:
  - url: http://localhost:8080
    description: Local / lab default
  - url: "{scheme}://{host}"
    description: Deployed gateway
    variables:
      scheme:
        default: https
...
paths:
...
  /api/v1/mfa/enroll:
    post:
      tags: [MFA]
      summary: Start TOTP enrollment
      responses:
        "200":
          description: Secret and backup codes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MFAEnrollResponse"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/mfa/confirm:
    post:
      tags: [MFA]
      summary: Confirm TOTP enrollment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
      responses:
        "200":
          description: MFA enabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  mfa_enabled: { type: boolean }
  /api/v1/mfa/disable:
    post:
      tags: [MFA]
      summary: Disable MFA
      description: Rejected with 409 while a password is set — clear the password first.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, description: TOTP or backup code }
      responses:
        "200":
          description: MFA disabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  mfa_enabled: { type: boolean }
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/auth/password:
    post:
      tags: [Auth]
      summary: Set or change account password
      description: |
        Always requires a TOTP code. If MFA is not enabled yet, start `POST /mfa/enroll`
        first; this call confirms enrollment atomically with the password set.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password, totp_code]
              properties:
                password: { type: string, format: password, minLength: 10 }
                totp_code: { type: string }
      responses:
        "200":
          description: Password saved; MFA enabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  password_set: { type: boolean }
                  mfa_enabled: { type: boolean }
                  must_set_password: { type: boolean }
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [Auth]
      summary: Clear account password
      description: Requires a TOTP code. MFA may remain for other login paths.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [totp_code]
              properties:
                totp_code: { type: string }
      responses:
        "200":
          description: Password cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  password_set: { type: boolean }
                  must_set_password: { type: boolean }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    SessionHeader:
      type: apiKey
      in: header
      name: X-Session-Token
    SessionCookie:
      type: apiKey
      in: cookie
      name: session_token
    SessionQuery:
      type: apiKey
      in: query
      name: token
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    MFAEnrollResponse:
      type: object
      properties:
        secret: { type: string }
        otpauth_url: { type: string }
        backup_codes:
          type: array
          items: { type: string }
        message: { type: string }
...
security:
  - ApiKeyAuth: []
  - SessionHeader: []
  - SessionCookie: []
  - BearerJWT: []